# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API для работы с вопросами",
        "title": "Questions API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
        "/options/add-option": {
            "post": {
                "description": "Создает новый вариант ответа для вопроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Добавить новый вариант ответа",
                "parameters": [
                    {
                        "description": "Объект варианта ответа",
                        "name": "option",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mobqom_questions_internal_dto.AddOptionDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Options"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/options/find-all": {
            "get": {
                "description": "Возвращает список всех вариантов ответов из базы данных",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Получить все варианты ответов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Options"
                            }
                        }
                    }
                }
            }
        },
        "/options/{questionId}": {
            "get": {
                "description": "Возвращает список всех вариантов ответов для конкретного вопроса",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "options"
                ],
                "summary": "Получить варианты ответов для вопроса",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID вопроса",
                        "name": "questionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Options"
                            }
                        }
                    }
                }
            }
        },
        "/questions/add-question": {
            "post": {
                "description": "Создает новый вопрос в базе данных",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Добавить новый вопрос",
                "parameters": [
                    {
                        "description": "Объект вопроса",
                        "name": "question",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_mobqom_questions_internal_dto.AddQuestionDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Question"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/questions/find-all": {
            "get": {
                "description": "Возвращает список всех вопросов из базы данных",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Получить все вопросы",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Question"
                            }
                        }
                    }
                }
            }
        },
        "/questions/find-by-game": {
            "get": {
                "description": "Возвращает список всех вопросов для конкретной игры",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Получить вопросы по ID игры",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "gameId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Question"
                            }
                        }
                    }
                }
            }
        },
        "/questions/random": {
            "get": {
                "description": "Возвращает случайный вопрос из базы данных",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "Получить случайный вопрос",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID игры",
                        "name": "gameId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Question"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_mobqom_questions_internal_domain.Options": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "question_id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_mobqom_questions_internal_domain.Question": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "game": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_mobqom_questions_internal_domain.Options"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_mobqom_questions_internal_dto.AddOptionDto": {
            "type": "object",
            "required": [
                "content",
                "question_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "question_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_mobqom_questions_internal_dto.AddQuestionDto": {
            "type": "object",
            "required": [
                "content",
                "game"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "game": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    }
}